feat(positions): add API for listing positions
Added a new API endpoint to retrieve a list of positions. This includes
the creation of the `PositionListAPIView` class and the corresponding
selector method `get_positions()` to fetch positions from the database.
Updated the URL patterns to include a new route for this endpoint.

The new endpoint allows clients to obtain a list of available positions
in the system.